# assiduidade_system_backend/openapi.yaml
openapi: 3.0.3
info:
  title: FaceClock API
  version: 1.0.0
  description: API principal do MVP de registro de ponto com reconhecimento facial.
servers:
  - url: https://api.faceclock.local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
        code:
          type: string
...
    ClockRecord:
      type: object
      required: [id, user_id, event_type, recorded_at, source, sync_status]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        device_id:
          type: string
          format: uuid
        event_type:
          type: string
          enum: [ENTRY, BREAK_START, BREAK_END, EXIT]
        recorded_at:
          type: string
          format: date-time
        source:
          type: string
          enum: [ONLINE, OFFLINE_SYNC, MANUAL, INTEGRATION]
        sync_status:
          type: string
          enum: [SYNCED, PENDING, FAILED]
        confidence_score:
          type: number
          format: double
        liveness_score:
          type: number
          format: double
        created_at:
          type: string
          format: date-time
...
    AdjustmentReviewRequest:
      type: object
      required: [status, review_notes]
      properties:
        status:
          type: string
          enum: [APPROVED, REJECTED]
        review_notes:
          type: string
    AdjustmentRequest:
      type: object
      required: [id, user_id, status, reason, created_at]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        clock_record_id:
          type: string
          format: uuid
        requested_event_type:
          type: string
          enum: [ENTRY, BREAK_START, BREAK_END, EXIT]
        requested_recorded_at:
          type: string
          format: date-time
        reason:
          type: string
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED, CANCELLED]
        review_notes:
          type: string
        reviewer_id:
          type: string
          format: uuid
        reviewed_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    PaginatedClockRecords:
      type: object
      required: [items, page, page_size, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ClockRecord'
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
    PaginatedAdjustmentRequests:
      type: object
      required: [items, page, page_size, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdjustmentRequest'
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
    ExportResponse:
      type: object
      required: [download_url, format]
      properties:
        download_url:
          type: string
          format: uri
        format:
          type: string
          enum: [csv]
    AuditLog:
      type: object
      required: [id, actor_type, action, entity_type, payload_hash, created_at]
      properties:
        id:
          type: string
          format: uuid
        actor_type:
          type: string
        actor_id:
          type: string
          format: uuid
        action:
          type: string
        entity_type:
          type: string
        entity_id:
          type: string
          format: uuid
        payload_hash:
          type: string
        previous_hash:
          type: string
        created_at:
          type: string
          format: date-time
    PaginatedAuditLogs:
      type: object
      required: [items, page, page_size, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AuditLog'
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
...
paths:
...
  /api/v1/admin/clock-records:
    get:
      tags: [Admin]
      security:
        - bearerAuth: []
      summary: Consulta administrativa de registros
      parameters:
        - in: query
          name: user_id
          schema:
            type: string
            format: uuid
        - in: query
          name: unit_id
          schema:
            type: string
            format: uuid
        - in: query
          name: event_type
          schema:
            type: string
            enum: [ENTRY, BREAK_START, BREAK_END, EXIT]
        - in: query
          name: start_date
          schema:
            type: string
            format: date
        - in: query
          name: end_date
          schema:
            type: string
            format: date
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: page_size
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Lista paginada de registros
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedClockRecords'
  /api/v1/admin/reports/export:
    get:
      tags: [Admin]
      security:
        - bearerAuth: []
      summary: Exporta relatorio administrativo basico
      parameters:
        - in: query
          name: unit_id
          schema:
            type: string
            format: uuid
        - in: query
          name: start_date
          schema:
            type: string
            format: date
        - in: query
          name: end_date
          schema:
            type: string
            format: date
        - in: query
          name: format
          schema:
            type: string
            enum: [csv]
      responses:
        '200':
          description: Link de exportacao gerado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportResponse'
  /api/v1/admin/reports/export.csv:
    get:
      tags: [Admin]
      security:
        - bearerAuth: []
      summary: Exporta relatorio administrativo em CSV
      parameters:
        - in: query
          name: user_id
          schema:
            type: string
            format: uuid
        - in: query
          name: unit_id
          schema:
            type: string
            format: uuid
        - in: query
          name: event_type
          schema:
            type: string
            enum: [ENTRY, BREAK_START, BREAK_END, EXIT]
        - in: query
          name: start_date
          schema:
            type: string
            format: date
        - in: query
          name: end_date
          schema:
            type: string
            format: date
      responses:
        '200':
          description: CSV exportado
          content:
            text/csv:
              schema:
                type: string
  /api/v1/admin/adjustments:
    get:
      tags: [Admin]
      security:
        - bearerAuth: []
      summary: Lista solicitacoes de ajuste
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [PENDING, APPROVED, REJECTED, CANCELLED]
        - in: query
          name: user_id
          schema:
            type: string
            format: uuid
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: page_size
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Lista paginada de ajustes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAdjustmentRequests'
  /api/v1/admin/adjustments/{adjustment_id}:
    patch:
      tags: [Admin]
      security:
        - bearerAuth: []
      summary: Aprova ou rejeita ajuste de ponto
      parameters:
        - in: path
          name: adjustment_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdjustmentReviewRequest'
      responses:
        '200':
          description: Ajuste revisado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdjustmentRequest'
        '409':
          description: Solicitacao ja revisada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Solicitacao nao encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/audit/logs:
    get:
      tags: [Audit]
      security:
        - bearerAuth: []
      summary: Consulta logs de auditoria
      parameters:
        - in: query
          name: entity_type
          schema:
            type: string
        - in: query
          name: entity_id
          schema:
            type: string
            format: uuid
        - in: query
          name: action
          schema:
            type: string
        - in: query
          name: start_date
          schema:
            type: string
            format: date
        - in: query
          name: end_date
          schema:
            type: string
            format: date
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: page_size
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Lista paginada de logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAuditLogs'
...